replaced success and confirm dialogs
